+++ src/grpc/proto/pulumirpc/pulumirpc.jl
@@ -4,9 +4,9 @@ include("../google/google.jl")
 import .pulumirpc as var"#pulumirpc"
 
 include("plugin_pb.jl")
+include("alias_pb.jl")
 include("source_pb.jl")
 include("engine_pb.jl")
-include("alias_pb.jl")
 include("callback_pb.jl")
 include("provider_pb.jl")
 include("codegen/codegen.jl")